+++ gql/queries/FTVACollectionDetail.gql
@@ -2,100 +2,29 @@
 
 query FTVACollectionDetail($slug: [String!]) {
   ftvaCollection: entry(section: "ftvaCollection", slug: $slug) {
-    id
     typeHandle
+    id
+    title
+    sectionTitle
     slug
-    uri
-
-    imageCarousel {
-      ...on imageCarousel_imageCarousel_BlockType {
-        image {
-          ...Image
-        }
-        creditText
-      }
+    to: uri
+    ftvaHomepageTitle
+    ftvaHomepageDescription
+    ftvaCollectionType
+    richText 
+    videoEmbed
+    viewAllSectionLink
+    infoBlock 
+    heroimage: ftvaHeroImage {
+     ...Image
     }
+    # image: ftvaImage {
+    #  ...Image
+    # }
 
-    # CardMeta
-    # title: eventTitle
-    # guestSpeaker
-    # tagLabels: ftvaEventFilters {
+    # ftvaRelatedCollections {
     #   title
     #   uri
-    # }
-    # introduction: ftvaEventIntroduction
-
-    eventDescription
-    # acknowledements: richText
-
-    # ftvaEventScreeningDetails {
-    #   ...on ftvaEventScreeningDetails_screeningDetails_BlockType {
-    #     title: screeningTitle
-    #     alternateTitle
-    #     language: languageTranslated
-    #     year
-    #     country
-    #     languageInfo
-    #     runtime
-    #     tagLabels: screeningTags {
-    #       title
-    #     }
-    #     text: descriptionOfScreening
-    #     trailer
-    #     image {
-    #       ...on mediaAndDocuments_Asset {
-    #         ...Image
-    #       }
-    #     }
-    #   }
-    # }
-
-    # SIDEBAR
-    # BlockEventDetail
-    startDateWithTime
-    # startDate: startDateWithTime @formatDateTime(format: "Y-m-d", timezone: "America/Los_Angeles")
-    # startTime: startDateWithTime @formatDateTime(format: "TH:i:s", timezone: "America/Los_Angeles")
-    # location {
-    #   title
-    #   campusMapId
-    #   publicUrl
-    #   address
-    # }
-
-    # BlockInfo / Ticket Info
-    # ftvaTicketInformation {
-    #   title
-    #   uri
-    # }
-    # ftvaEventRegistration {
-    #   ...on ftvaEventRegistration_eventRegistration_BlockType {
-    #     registrationText
-    #     registrationUrlLink
-    #   }
     # }
   }
-
-  # SectionTeaserCard
-  # ftvaEventSeries: entries(
-  #   section: "ftvaEventSeries",  orderBy: "postDate",
-  #   relatedToEntries: {section: "ftvaEvent", slug: $slug}) {
-  #   ... on ftvaEventSeries_eventSeries_Entry {
-  #     id
-  #     slug
-  #     typeHandle
-  #     title
-  #     ftvaEvent(startDateWithTime: ">= now") {
-  #       ... on ftvaEvent_ftvaEvent_Entry {
-  #         id
-  #         to: slug
-  #         title: eventTitle
-  #         startDate: startDateWithTime @formatDateTime(format: "Y-m-d", timezone: "America/Los_Angeles")
-  #         startTime: startDateWithTime @formatDateTime(format: "TH:i:s", timezone: "America/Los_Angeles")
-  #         image: ftvaImage {
-  #           ...Image
-  #         }
-  #       }
-  #     }
-  #   }
-  # }
 }
